Added a comment feature to project.
Added a couple of database migrations to add a comments resource. The first one "CommentCreate" didn't seem to work, and I messed up my migrations locally by trying to modify the migration. So instead, I made a new one that seemed to work. After that I added the models, views and serializers needed to view the list of comments and create new ones. I then added a specific URL for the comments. This seems to work. I think that the "project" field is supposed to be a reference to the ID of an existing project but I didn't get that far